% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP.R
\name{count.xena}
\alias{count.xena}
\title{Count unique occurrences of a feature in XENA object.}
\usage{
\method{count}{xena}(x, ft, sort = FALSE, name = NULL)
}
\arguments{
\item{x}{a XENA class object.}

\item{ft}{a variable to count.}

\item{sort}{logical, if TRUE, will show the largest groups at the top.}

\item{name}{The name of the new column in the output.
If omitted, it will default to n.}
}
\value{
a data frame with counts, frequencies and percentages for each
category of a variable.
}
\description{
Counts occurrences of unique categories of a variable in
a XENA object.
}
\details{
Functions in a similar way as the genuine `count()` function from
the dplyr package but handles also 'fuzzy' categories such as mutation
effects with multiple categories per sample seperated by semicolon.
A sample is counted once even if it e.g. has two missense mutations.
}
