% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{count_mutations}
\alias{count_mutations}
\title{Count mutations per sample.}
\usage{
count_mutations(xena_db, ...)
}
\arguments{
\item{xena_db}{a XENA database object.}

\item{...}{extra arguments, currently none.}
}
\value{
a data frame listing the samples and the number of mutations per
sample.
}
\description{
Counts all mutations per sample provided in the XENA class
database.
}
