% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{freq_mutations}
\alias{freq_mutations}
\title{Gen frequency of mutated genes.}
\usage{
freq_mutations(xena_db, .parallel = FALSE, ...)
}
\arguments{
\item{xena_db}{a XENA database object.}

\item{.parallel}{logical, should the computation be run in parallel?}

\item{...}{extra arguments, currently none.}
}
\value{
a data frame with the frequency and percent of mutations.
}
\description{
Calculates frequencies of the mutated genes in the database.
}
