% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP.R
\name{is_xena}
\alias{is_xena}
\title{Cehcks for the XENA class.}
\usage{
is_xena(x)
}
\arguments{
\item{x}{an object.}
}
\value{
a logical value.
}
\description{
Checks of the object is an instance of the XENA class.
}
