% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_xena}
\alias{load_xena}
\title{Load a XENA data base.}
\usage{
load_xena(path)
}
\arguments{
\item{path}{path to the XENA database file; accepts .tar or .gz archives.}
}
\value{
a XENA object, with a data frame interface.
}
\description{
Loads a XENA database from the disc and converts it to a
tidyverse - compatible format.
}
\details{
The XENA object enhances functionality of tibbles to hold
mutation data. You may manipulate the XENA object with a plethora of
tidyverse tools in a very similar way as an ordinary tibble.
}
