% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP.R
\name{search.xena}
\alias{search.xena}
\title{Query a local XENA database.}
\usage{
\method{search}{xena}(
  x,
  key = c("gene", "chrom", "effect", "mutation_id"),
  value = NULL,
  regex = NULL,
  unique_samples = FALSE,
  join = FALSE,
  detailed = FALSE
)
}
\arguments{
\item{x}{a XENA object.}

\item{key}{name of the database field to search in. Possible values are:
'gene' (gene with the mutation), 'chrom' (chromosome), 'effect' (effect of
the mutation), 'mutation_id' (mutation ID).}

\item{value}{a key value or a vector for the direct search.}

\item{regex}{a regular expression to scan the key, ignored if value is
provided.}

\item{unique_samples}{logical, should duplicated samples be excluded from
the output?}

\item{join}{logical, should the output be merged into a single data frame by
the sample ID?}

\item{detailed}{logical, if TRUE a XENA object being a subset of the input
is returned.}
}
\value{
a data frame or a XENA object, if detailed = TRUE.
}
\description{
Searches for samples carrying the specified mutation.
Lists all the samples with a mutation of the given gene, chromosome, with
the effect provided or mutation ID. Searches directly via the key value or
with a regular expression. The option 'detailed' returns a complete subset
of the xena_mut object. Otherwise, only the sample ids and vaf value are
returned.
}
