% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tab_mutations}
\alias{tab_mutations}
\title{List mutated genes.}
\usage{
tab_mutations(xena_db, as_matrix = TRUE, sparse = NULL, .parallel = FALSE, ...)
}
\arguments{
\item{xena_db}{a XENA database object.}

\item{as_matrix}{logical, should the output be in a matrix form?}

\item{sparse}{logical or NULL, specifying if the result should be sparse or
not. By default, it is made sparse when more than half of the entries are 0.}

\item{.parallel}{logical, should the operation be run in parallel? Runs via
the foreach backend, the user need to register the process prior to calling
the function.
If FALSE, a data.frame is returned (slower).}

\item{...}{extra arguments, currently none.}
}
\value{
a (sparse) matrix or a data frame.
}
\description{
Creates a matrix or table with the mutated genes for each
sample (0: wt, 1: a mutation listed in the database). Specificity towards
specific mutations (e.g. missense ones) and their quantity (Variant Allele
Frequency) can be acheived by simple filtering of the input xena_bd object
(inherits the transformation interface from data.frame).
}
